% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison_metrics.R
\name{comparison_metrics}
\alias{comparison_metrics}
\title{Function comparing GAMs by different metrics}
\usage{
comparison_metrics(gam_comb_cond, gam_comb_no_cond)
}
\arguments{
\item{gam_comb_cond}{Model which includes condition as a factor}

\item{gam_comb_no_cond}{Model which does not include condition as a factor}
}
\value{
A data frame with different metrics by which one can compare two different models
}
\description{
Function comparing GAMs by different metrics
}
