% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_array.R
\name{LRT_array}
\alias{LRT_array}
\title{Function that creates an array for SPM map}
\usage{
LRT_array(data_gz, cond_1_txt, cond_2_txt)
}
\arguments{
\item{data_gz}{Main data file (.gz extension) with BOLD signal values}

\item{cond_1_txt}{The file stating when stimuli 1 occurred}

\item{cond_2_txt}{The file stating when stimuli 2 occurred}
}
\value{
The array of p-values that can be used to create a SPM map
}
\description{
Function that creates an array for SPM map
}
