% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_comparison.R
\name{LRT_comparison}
\alias{LRT_comparison}
\title{Function producing p-values that help with models assessment}
\usage{
LRT_comparison(gam_comb_cond, gam_comb_no_cond)
}
\arguments{
\item{gam_comb_cond}{Model which includes condition as a factor}

\item{gam_comb_no_cond}{Model which does not include condition as a factor}
}
\value{
A p-value corresponding to the Likelihood Ratio Test
}
\description{
Function producing p-values that help with models assessment
}
