% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_and_plot_gam.R
\name{predict_and_plot_gam}
\alias{predict_and_plot_gam}
\title{Function predicting the interpolated values of fMRI}
\usage{
predict_and_plot_gam(
  data_con_1,
  data_con_2,
  data_comb,
  gam_con_1,
  gam_con_2,
  gam_comb_cond,
  gam_comb_no_cond,
  no_of_inter = 100,
  CI = 0.05
)
}
\arguments{
\item{data_con_1}{Data frame for first conditions}

\item{data_con_2}{Data frame for second conditions}

\item{data_comb}{Data frame for combined conditions}

\item{gam_con_1}{Model which includes first condition only}

\item{gam_con_2}{Model which includes second condition only}

\item{gam_comb_cond}{Model which includes condition as a factor}

\item{gam_comb_no_cond}{Model which does not include condition as a factor}

\item{no_of_inter}{Number of interpolated values}

\item{CI}{Confidence interval for the plot drawing}
}
\value{
A matrix with predictions for both models.
}
\description{
Function predicting the interpolated values of fMRI
}
