% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gam.R
\name{predict_gam}
\alias{predict_gam}
\title{Function predicting the interpolated values of fMRI}
\usage{
predict_gam(data_comb, gam_comb_cond, gam_comb_no_cond, no_of_inter = 100)
}
\arguments{
\item{data_comb}{Data frame for combined conditions}

\item{gam_comb_cond}{Model which includes condition as a factor}

\item{gam_comb_no_cond}{Model which does not include condition as a factor}

\item{no_of_inter}{Number of interpolated values}
}
\value{
A matrix with predictions for both models.
}
\description{
Function predicting the interpolated values of fMRI
}
