% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function preparing a data frame for the further analysis}
\usage{
prepare_data(data_gz, cond_1_txt, cond_2_txt, x, y, z, num_post_TR = 10)
}
\arguments{
\item{data_gz}{Main data file (.gz extension) with BOLD signal values}

\item{cond_1_txt}{The file stating when stimuli 1 occurred}

\item{cond_2_txt}{The file stating when stimuli 2 occurred}

\item{x}{First dimension}

\item{y}{Second dimension}

\item{z}{Third dimension}
}
\value{
A polished data set prepared for modelling
}
\description{
Function preparing a data frame for the further analysis
}
