% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Function fitting the GAMs to previously pre-processed data.}
\usage{
fit_gam(
  data_comb,
  smoothing_basis = "tp",
  number_of_knots = 10,
  estimating_method = "REML"
)
}
\arguments{
\item{data_comb}{Data frame for combined conditions}

\item{smoothing_basis}{Smooth term}

\item{number_of_knots}{Basis dimension choice for smooths}

\item{estimating_method}{The smoothing parameter estimation method}
}
\value{
List of 2 GAM models. One including condition as a factor and one not.
}
\description{
Function fitting the GAMs to previously pre-processed data.
}
