% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Function for modelling fMRI data}
\usage{
fit_gam(
  data_frame,
  smoothing_basis = "tp",
  number_of_knots = 10,
  method = "REML"
)
}
\arguments{
\item{data_frame}{A processed data frame}

\item{smoothing_basis}{A smoothing function used for modelling}

\item{number_of_knots}{A number of knots used for local smoothing}

\item{method}{A method of estimation}
}
\value{
A mdoel
}
\description{
Function for modelling fMRI data
}
