% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Function fitting the GAMs to previously pre-processed data.}
\usage{
fit_gam(
  data_con_1,
  data_con_2,
  data_comb,
  smoothing_basis = "tp",
  number_of_knots = 10,
  estimating_method = "REML"
)
}
\arguments{
\item{data_con_1}{Data frame for condition 1}

\item{data_con_2}{Data frame for condition 2}

\item{data_comb}{Data frame for combined conditions}

\item{smoothing_basis}{Smooth term}

\item{number_of_knots}{Basis dimension choice for smooths}

\item{estimating_method}{The smoothing parameter estimation method}
}
\value{
List of 3 GAM models, one for the first condition, one for second, and one for combined case scenario.
}
\description{
Function fitting the GAMs to previously pre-processed data.
}
